requirements.txt
feedparser>=6.0
beautifulsoup4>=4.12
httpx>=0.27
aiohttp>=3.9
python-dateutil>=2.9
